# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Availo Ventry - Smart Visitor Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout current user
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/accept-invitation:
    post:
      operationId: acceptInvitation
      tags: [auth]
      summary: Accept an invitation and set up account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInvitationRequest"
      responses:
        "200":
          description: Invitation accepted, account set up
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-invitation-token:
    get:
      operationId: verifyInvitationToken
      tags: [auth]
      summary: Verify invitation token validity
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationTokenInfo"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change user password
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── ORGANIZATIONS ───────────────────────────────────────────────────────────
...
  /organizations/{orgId}/blacklist:
    get:
      operationId: listBlacklist
      tags: [blacklist]
      summary: List blacklisted visitors
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of blacklisted visitors
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedBlacklist"
    post:
      operationId: addToBlacklist
      tags: [blacklist]
      summary: Add visitor to blacklist
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToBlacklistRequest"
      responses:
        "201":
          description: Added to blacklist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlacklistEntry"
  /organizations/{orgId}/blacklist/{blacklistId}:
    delete:
      operationId: removeFromBlacklist
      tags: [blacklist]
      summary: Remove visitor from blacklist
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
        - name: blacklistId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed from blacklist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── DASHBOARD ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  parameters:
    orgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - total
        - page
        - limit
        - totalPages

    # ── Enums ──────────────────────────────────────────────────────────────────
    UserRole:
      type: string
      enum:
        - super_admin
        - org_admin
        - visitor_manager
        - receptionist
        - host_employee
    OrgStatus:
      type: string
      enum:
        - active
        - suspended
        - pending_setup
        - deactivated
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/CurrentUser"
        token:
          type: string
      required:
        - user
    CurrentUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        nameAr:
          type: string
          nullable: true
        role:
          $ref: "#/components/schemas/UserRole"
        orgId:
          type: string
          nullable: true
        branchId:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        jobTitle:
          type: string
          nullable: true
        mustChangePassword:
          type: boolean
        setupWizardCompleted:
          type: boolean
        organizationName:
          type: string
          nullable: true
        organizationStatus:
          $ref: "#/components/schemas/OrgStatus"
          nullable: true
      required:
        - id
        - email
        - name
        - role
        - mustChangePassword
        - setupWizardCompleted
    AcceptInvitationRequest:
      type: object
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
        twoFactorEnabled:
          type: boolean
        acceptedTerms:
          type: boolean
      required:
        - token
        - password
        - acceptedTerms
    InvitationTokenInfo:
      type: object
      properties:
        valid:
          type: boolean
        email:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        orgName:
          type: string
        expiresAt:
          type: string
          format: date-time
      required:
        - valid
        - email
        - name
        - role
        - orgName
        - expiresAt
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
      required:
        - currentPassword
        - newPassword

    # ── Organizations ─────────────────────────────────────────────────────────
...
    Visitor:
      type: object
      properties:
        id:
          type: string
        nationalIdNumber:
          type: string
          nullable: true
        fullName:
          type: string
        fullNameAr:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        companyName:
          type: string
          nullable: true
        photoUrl:
          type: string
          nullable: true
        isBlacklisted:
          type: boolean
        verificationStatus:
          type: string
          enum:
            - pending
            - verified_otp
            - verified_nafath
            - verified_both
            - skipped_trusted
            - unverified
        lastVerifiedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - isBlacklisted
        - verificationStatus
        - createdAt
...
    BlacklistEntry:
      type: object
      properties:
        id:
          type: string
        orgId:
          type: string
        visitorId:
          type: string
        visitor:
          $ref: "#/components/schemas/Visitor"
          nullable: true
        reason:
          type: string
        blacklistedById:
          type: string
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orgId
        - visitorId
        - reason
        - blacklistedById
        - createdAt
    AddToBlacklistRequest:
      type: object
      properties:
        visitorId:
          type: string
        reason:
          type: string
        expiresAt:
          type: string
          format: date-time
          nullable: true
      required:
        - visitorId
        - reason
    PaginatedBlacklist:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/BlacklistEntry"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta

    # ── Dashboard ─────────────────────────────────────────────────────────────
...